Changed update, delete, and search functionality
Because of conflicting messages on Piazza, I've changed the output of
update to show the total number of items of the same name in the cart,
so if there are 2 items labeled PS3, for example, at different or same
prices, an update method will reflect the total number of items
currently in the cart with that same name with the new quantity added.

Same for delete, the sum of the number of items labeled PS3, if there
are multiple, will now be output instead of the number of Item objects
in the cart that have the name PS3. Same for search as well, outputting
the quantity of all objects with the name PS3, for example, not just the
number of Item objects in the shopping cart that are called PS3.